pyproject.toml
[project]
name = "gcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-settings>=2.0.0",
    "google-adk>=1.27.2",
    "asyncpg>=0.30.0",
    "openinference-instrumentation-google-adk>=0.1.10",
    "arize-phoenix-otel>=0.15.0",
]
